// server/src/payment_methods/dto/momo-success.dto.ts
export class MomoSuccessDto {
  deposit_id: string;
  partnerCode: string;
  orderId: string;
  requestId: string;
  amount: string;
  orderInfo: string;
  orderType: string;
  transId: string;
  resultCode: string;
  message: string;
  payType: string;
  responseTime: string;
  extraData: string;
  signature: string;
}